conrad_backend/requirements.txt
fastapi
uvicorn[standard]
requests
google-generativeai
python-dotenv
atlassian-python-api
pydantic
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
nltk
scikit-learn
